+++ src-plugin/Plugin/PluginEvents.cs
@@ -98,6 +98,7 @@ namespace K4Arenas
 
 				WaitingArenaPlayers = new Queue<ArenaPlayer>(WaitingArenaPlayers.Where(p => p.Controller != playerController));
 				Arenas?.ArenaList.ForEach(arena => arena.RemovePlayer(playerController));
+				TerminateRoundIfPossible();
 				return HookResult.Continue;
 			});
 
@@ -380,7 +381,7 @@ namespace K4Arenas
 				}
 
 				return HookResult.Continue;
-			});
+			}, HookMode.Pre);
 
 			RegisterEventHandler((EventRoundStart @event, GameEventInfo info) =>
 			{