+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+-- 2024.12.23 - 1.5.5
+
+- feat: Added 3 new natives - IsAFK, FindOpponents & TerminateRoundIfPossible
+- update: When a player executes !afk, execute TerminateRoundIfPossible.
+- update: Added `allowed-weapon-prefs` to toggle showing weaponTypes in `!guns` menu.
+
 -- 2024.12.20 - 1.5.4
 
 - fix: PerformAFKAction didnt change the AFK state of player